backend/requirements.txt
fastapi
httpx
python-dotenv
crewai
uvicorn
